+++ src/test/functionalTests/tests/prod/LRvLip_response_admitAll_pininpost_tests.js
@@ -41,4 +41,4 @@ Scenario('Response with AdmitAll and Immediate payment @citizenUI @admitAll @nig
   // await api.enterBreathingSpace(config.applicantSolicitorUser);
   // await api.liftBreathingSpace(config.applicantSolicitorUser);
   await api.viewAndRespondToDefence(config.applicantSolicitorUser, config.defenceType.admitAllPayImmediate);
-}).tag('@regression-cui-r1');
+});
